# api/openapi.yaml
openapi: 3.0.3
info:
  title: RideLite API
  version: 1.0.0
  description: |
    REST API for the RideLite mobility platform exposing trip management,
    ETA estimation and observability endpoints through the API gateway.
servers:
  - url: http://localhost:8088
    description: Local development gateway
paths:
  /v1/trips:
    post:
      summary: Create a new trip request
      description: |
        Creates a new trip request for a rider and attempts to reserve a driver via the
        matching engine. Idempotent when an `Idempotency-Key` header is provided.
      operationId: createTrip
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
          description: Unique key that guarantees idempotent processing of the request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTripRequest'
            examples:
              standard:
                summary: Basic request
                value:
                  rider_id: "11111111-1111-4111-8111-111111111111"
                  pickup:
                    lat: 35.741
                    lng: 51.421
                  dropoff:
                    lat: 35.752
                    lng: 51.444
                  vehicle_type: economy
      responses:
        '201':
          description: Trip created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTripResponse'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/eta:
    get:
      summary: Estimate ETA
      description: Returns ETA estimations for the nearest driver and trip duration.
      operationId: getEta
      parameters:
        - in: query
          name: pickup_lat
          required: true
          schema:
            type: number
            format: float
        - in: query
          name: pickup_lng
          required: true
          schema:
            type: number
            format: float
        - in: query
          name: dropoff_lat
          required: true
          schema:
            type: number
            format: float
        - in: query
          name: dropoff_lng
          required: true
          schema:
            type: number
            format: float
      responses:
        '200':
          description: Estimated durations in seconds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ETAResponse'
...
components:
...
  schemas:
    GeoPoint:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          format: float
        lng:
          type: number
          format: float
...
    CreateTripRequest:
      type: object
      required: [rider_id, pickup, dropoff]
      properties:
        rider_id:
          type: string
          format: uuid
        pickup:
          $ref: '#/components/schemas/GeoPoint'
        dropoff:
          $ref: '#/components/schemas/GeoPoint'
        vehicle_type:
          type: string
    CreateTripResponse:
      type: object
      properties:
        trip_id:
          type: string
          format: uuid
        status:
          type: string
      required: [trip_id, status]
...
    ETAResponse:
      type: object
      properties:
        driver_eta_sec:
          type: number
          format: float
        trip_eta_sec:
          type: number
          format: float
        driver_id:
          type: string
          format: uuid
          nullable: true
      required:
        - driver_eta_sec
        - trip_eta_sec
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error